rename flightplandb file, fix __all__ in __init__.py
