+++ src/flightplandb/__init__.py
@@ -3,5 +3,8 @@
 # Version of the flightplandb package
 __version__ = "0.5.0"
 
-from flightplandb.datatypes import *  # noqa: F403, F401
-from flightplandb.submodules import *  # noqa: F403, F401
+from . import internal, exceptions, datatypes, submodules
+
+# from flightplandb.datatypes import *  # noqa: F403, F401
+# from flightplandb.submodules import *  # noqa: F403, F401
+__all__ = ["internal", "exceptions", "datatypes", "submodules"]
