+++ src/flightplandb/submodules/__init__.py
@@ -1,1 +1,2 @@
+from . import api, nav, plan, tags, user, weather
 __all__ = ["api", "nav", "plan", "tags", "user", "weather"]
